Remove BicepFile.Dummy second iteration
