+++ src/Bicep.Core/Registry/Oci/IOciArtifactAddressComponents.cs
@@ -7,7 +7,7 @@ namespace Bicep.Core.Registry.Oci
     /// Represents a module/provider reference without the scheme ("br:") and without any aliases
     /// </summary>
     /// <example>test.azurecr.io/foo/bar:latest</example>
-    public interface IArtifactAddressComponents
+    public interface IOciArtifactAddressComponents
     {
         /// <summary>
         /// Gets the registry URI.
