+++ src/Bicep.Core/Registry/Oci/OciArtifactReference.cs
@@ -15,7 +15,7 @@ namespace Bicep.Core.Registry.Oci
 {
     public class OciArtifactReference : ArtifactReference, IOciArtifactReference
     {
-        public OciArtifactReference(BicepSourceFile referencingFile, ArtifactType type, IArtifactAddressComponents artifactIdParts) :
+        public OciArtifactReference(BicepSourceFile referencingFile, ArtifactType type, IOciArtifactAddressComponents artifactIdParts) :
             base(referencingFile, OciArtifactReferenceFacts.Scheme)
         {
             Type = type;
@@ -34,10 +34,10 @@ namespace Bicep.Core.Registry.Oci
             }
 
             Type = type;
-            AddressComponents = new ArtifactAddressComponents(registry, repository, tag, digest);
+            AddressComponents = new OciArtifactAddressComponents(registry, repository, tag, digest);
         }
 
-        public IArtifactAddressComponents AddressComponents { get; }
+        public IOciArtifactAddressComponents AddressComponents { get; }
 
         /// <summary>
         /// Gets the type of artifact reference.
@@ -74,18 +74,12 @@ namespace Bicep.Core.Registry.Oci
         public override bool IsExternal => true;
 
         // unqualifiedReference is the reference without a scheme or alias, e.g. "example.azurecr.invalid/foo/bar:v3"
-        // The configuration and parentModuleUri are needed to resolve aliases and experimental features
-        public static ResultWithDiagnosticBuilder<OciArtifactReference> TryParseModuleAndAlias(BicepSourceFile referencingFile, string? aliasName, string unqualifiedReference)
-            => TryParse(referencingFile, ArtifactType.Module, aliasName, unqualifiedReference);
-
-        public static ResultWithDiagnosticBuilder<OciArtifactReference> TryParseModule(string unqualifiedReference)
-            => TryParse(BicepFile.Dummy, ArtifactType.Module, null, unqualifiedReference);
-
+        // The referencingFile is needed to resolve aliases and experimental features
         public static ResultWithDiagnosticBuilder<OciArtifactReference> TryParse(BicepSourceFile referencingFile, ArtifactType type, string? aliasName, string unqualifiedReference)
         {
-            if (TryParseParts(referencingFile, type, aliasName, unqualifiedReference).IsSuccess(out var parts, out var errorBuilder))
+            if (TryParseComponents(referencingFile, type, aliasName, unqualifiedReference).IsSuccess(out var components, out var errorBuilder))
             {
-                return new(new OciArtifactReference(referencingFile, type, parts.Registry, parts.Repository, parts.Tag, parts.Digest));
+                return new(new OciArtifactReference(referencingFile, type, components.Registry, components.Repository, components.Tag, components.Digest));
             }
             else
             {
@@ -93,19 +87,8 @@ namespace Bicep.Core.Registry.Oci
             }
         }
 
-        // Doesn't handle aliases
-        public static ResultWithDiagnosticBuilder<IArtifactAddressComponents> TryParseFullyQualifiedComponents(string rawValue)
+        private static ResultWithDiagnosticBuilder<OciArtifactAddressComponents> TryParseComponents(BicepSourceFile referencingFile, ArtifactType type, string? aliasName, string unqualifiedReference)
         {
-            return TryParseParts(BicepFile.Dummy, ArtifactType.Module, aliasName: null, rawValue);
-        }
-
-        // TODO: Completely remove aliasName and configuration dependencies and move the non-dependent portion to a static method on ArtifactAddressComponents
-        private static ResultWithDiagnosticBuilder<IArtifactAddressComponents> TryParseParts(BicepSourceFile referencingFile, ArtifactType type, string? aliasName, string unqualifiedReference)
-        {
-            static string GetBadReference(string referenceValue) => $"{OciArtifactReferenceFacts.Scheme}:{referenceValue}";
-
-            static string DecodeSegment(string segment) => HttpUtility.UrlDecode(segment);
-
             if (aliasName is { })
             {
                 switch (type)
@@ -122,125 +105,7 @@ namespace Bicep.Core.Registry.Oci
                 }
             }
 
-            // the set of valid OCI artifact refs is a subset of the set of valid URIs if you remove the scheme portion from each URI
-            // manually prepending any valid URI scheme allows to get free validation via the built-in URI parser
-            if (!Uri.TryCreate($"{OciArtifactReferenceFacts.Scheme}://{unqualifiedReference}", UriKind.Absolute, out var artifactUri) ||
-                artifactUri.Segments.Length <= 1 ||
-                !string.Equals(artifactUri.Segments[0], "/", StringComparison.Ordinal))
-            {
-                return new(x => x.InvalidOciArtifactReference(aliasName, GetBadReference(unqualifiedReference)));
-            }
-
-            string registry = artifactUri.Authority;
-            if (registry.Length > OciArtifactReferenceFacts.MaxRegistryLength)
-            {
-                return new(x => x.InvalidOciArtifactReferenceRegistryTooLong(
-                    aliasName,
-                    GetBadReference(unqualifiedReference),
-                    registry,
-                    OciArtifactReferenceFacts.MaxRegistryLength));
-            }
-
-            // for "br://example.azurecr.io/foo/bar:v1", the segments are "/", "foo/", and "bar:v1"
-            // iterate only over middle segments
-            var repoBuilder = new StringBuilder();
-            for (int i = 1; i < artifactUri.Segments.Length - 1; i++)
-            {
-                // don't try to match the last character, which is always '/'
-                string segment = artifactUri.Segments[i];
-                var segmentWithoutTrailingSlash = segment[..^1];
-                if (!OciArtifactReferenceFacts.IsOciNamespaceSegment(segmentWithoutTrailingSlash))
-                {
-                    var invalidSegment = DecodeSegment(segmentWithoutTrailingSlash);
-                    return new(x => x.InvalidOciArtifactReferenceInvalidPathSegment(aliasName, GetBadReference(unqualifiedReference), invalidSegment));
-                }
-
-                // even though chars that require URL-escaping are not part of the allowed regexes
-                // users can still type them in, so error messages should contain the original text rather than an escaped version
-                repoBuilder.Append(DecodeSegment(segment));
-            }
-
-            // on a valid ref it would look something like "bar:v1" or "bar@sha256:e207a69d02b3de40d48ede9fd208d80441a9e590a83a0bc915d46244c03310d4"
-            var lastSegment = artifactUri.Segments[^1];
-
-            static (int index, char? delimiter) FindLastSegmentDelimiter(string lastSegment)
-            {
-                char[] delimiters = [':', '@'];
-                int index = lastSegment.IndexOfAny(delimiters);
-
-                return (index, index == -1 ? null : lastSegment[index]);
-            }
-
-            var (indexOfLastSegmentDelimiter, delimiter) = FindLastSegmentDelimiter(lastSegment);
-
-            // users will type references from left to right, so we should validate the last component of the module path
-            // before we complain about the missing tag, which is the last part of the module ref
-            var name = DecodeSegment(!delimiter.HasValue ? lastSegment : lastSegment[..indexOfLastSegmentDelimiter]);
-            if (!OciArtifactReferenceFacts.IsOciNamespaceSegment(name))
-            {
-                return new(x => x.InvalidOciArtifactReferenceInvalidPathSegment(aliasName, GetBadReference(unqualifiedReference), name));
-            }
-
-            repoBuilder.Append(name);
-
-            string repository = repoBuilder.ToString();
-            if (repository.Length > OciArtifactReferenceFacts.MaxRepositoryLength)
-            {
-                return new(x => x.InvalidOciArtifactReferenceRepositoryTooLong(
-                    aliasName,
-                    GetBadReference(unqualifiedReference),
-                    repository,
-                    OciArtifactReferenceFacts.MaxRepositoryLength));
-            }
-
-            // now we can complain about the missing tag or digest
-            if (!delimiter.HasValue)
-            {
-                return new(x => x.InvalidOciArtifactReferenceMissingTagOrDigest(aliasName, GetBadReference(unqualifiedReference)));
-            }
-
-            var tagOrDigest = DecodeSegment(lastSegment.Substring(indexOfLastSegmentDelimiter + 1));
-            if (string.IsNullOrEmpty(tagOrDigest))
-            {
-                return new(x => x.InvalidOciArtifactReferenceMissingTagOrDigest(aliasName, GetBadReference(unqualifiedReference)));
-            }
-
-            switch (delimiter.Value)
-            {
-                case ':':
-                    var tag = tagOrDigest;
-                    if (tag.Length > OciArtifactReferenceFacts.MaxTagLength)
-                    {
-                        return new(x => x.InvalidOciArtifactReferenceTagTooLong(
-                            aliasName,
-                            GetBadReference(unqualifiedReference),
-                            tag,
-                            OciArtifactReferenceFacts.MaxTagLength));
-                    }
-
-                    if (!OciArtifactReferenceFacts.IsOciTag(tag))
-                    {
-                        return new(x => x.InvalidOciArtifactReferenceInvalidTag(
-                            aliasName,
-                            GetBadReference(unqualifiedReference),
-                            tag));
-                    }
-
-                    return new(new ArtifactAddressComponents(registry, repository, tag, digest: null));
-
-                case '@':
-                    var digest = tagOrDigest;
-                    if (!OciArtifactReferenceFacts.IsOciDigest(digest))
-                    {
-                        return new(x => x.InvalidOciArtifactReferenceInvalidDigest(aliasName, GetBadReference(unqualifiedReference), digest));
-                    }
-
-                    return new(new ArtifactAddressComponents(registry, repository, tag: null, digest: digest));
-
-                default:
-                    throw new NotImplementedException($"Unexpected last segment delimiter character '{delimiter.Value}'.");
-            }
-
+            return OciArtifactAddressComponents.TryParse(unqualifiedReference, aliasName);
         }
 
         public override bool Equals(object? obj)
